GPUScreenGridLayer -> ScreenGridLayer, move GPUGridAggregator to Core as experimental feature
