+++ modules/core/src/index.js
@@ -80,6 +80,7 @@ import {default as BinSorter} from './utils/bin-sorter';
 import {defaultColorRange} from './utils/color-utils';
 import {linearScale, getLinearScale, quantizeScale, getQuantizeScale} from './utils/scale-utils';
 
+export {default as _GPUGridAggregator} from './experimental/utils/gpu-grid-aggregator';
 // Exports for layers
 // Experimental Features may change in minor version bumps, use at your own risk)
 export const experimental = {
